requirements.txt
SpeechRecognition
pyttsx3
PyAudio
python-dotenv
pywhatkit
duckduckgo-search
bs4
requests
edge-tts
pygame
google-genai
aiofiles
fastapi
uvicorn
psutil
